@@ -4,26 +4,21 @@
         "dev": "vite",
         "build": "vite build"
     },
+    "dependencies": {
+        "vite-plugin-vuetify": "^1.0.0",
+        "vue": "^3.2.45",
+        "vuetify": "^3.0.1"
+    },
     "devDependencies": {
         "@tailwindcss/forms": "^0.5.2",
+        "@vitejs/plugin-vue": "^3.2.0",
         "alpinejs": "^3.4.2",
         "autoprefixer": "^10.4.2",
         "axios": "^1.1.2",
-        "deepmerge": "^4.2.2",
         "laravel-vite-plugin": "^0.6.0",
         "lodash": "^4.17.19",
         "postcss": "^8.4.6",
-        "sass": "~1.32",
-        "sass-loader": "^13.2.0",
         "tailwindcss": "^3.1.0",
-        "unplugin-vue-components": "^0.22.9",
-        "vite": "^3.0.0",
-        "vite-plugin-vue2": "^2.0.2",
-        "vue-template-compiler": "^2.6.12"
-    },
-    "dependencies": {
-        "@vitejs/plugin-vue": "^2.3.3",
-        "vue": "^2.7.13",
-        "vuetify": "^2.6.12"
+        "vite": "^3.2.4"
     }
 }
