@@ -1,21 +1,17 @@
 <template>
   <v-app id="inspire">
-    <v-app-bar
-      app
-      color="white"
-      flat
-    >
-      <v-container class="py-0 fill-height">
+    <v-app-bar flat>
+      <v-container class="fill-height d-flex align-center">
         <v-avatar
-          class="mr-10"
-          color="grey darken-1"
+          class="mr-10 ml-4"
+          color="grey-darken-1"
           size="32"
         ></v-avatar>
 
         <v-btn
           v-for="link in links"
           :key="link"
-          text
+          variant="text"
         >
           {{ link }}
         </v-btn>
@@ -24,45 +20,39 @@
 
         <v-responsive max-width="260">
           <v-text-field
-            dense
-            flat
+            density="compact"
             hide-details
-            rounded
-            solo-inverted
+            variant="solo"
           ></v-text-field>
         </v-responsive>
       </v-container>
     </v-app-bar>
 
-    <v-main class="grey lighten-3">
+    <v-main class="bg-grey-lighten-3">
       <v-container>
         <v-row>
           <v-col cols="2">
             <v-sheet rounded="lg">
-              <v-list color="transparent">
+              <v-list rounded="lg">
                 <v-list-item
                   v-for="n in 5"
                   :key="n"
                   link
                 >
-                  <v-list-item-content>
-                    <v-list-item-title>
-                      List Item {{ n }}
-                    </v-list-item-title>
-                  </v-list-item-content>
+                  <v-list-item-title>
+                    List Item {{ n }}
+                  </v-list-item-title>
                 </v-list-item>
 
                 <v-divider class="my-2"></v-divider>
 
                 <v-list-item
                   link
-                  color="grey lighten-4"
+                  color="grey-lighten-4"
                 >
-                  <v-list-item-content>
-                    <v-list-item-title>
-                      Refresh
-                    </v-list-item-title>
-                  </v-list-item-content>
+                  <v-list-item-title>
+                    Refresh
+                  </v-list-item-title>
                 </v-list-item>
               </v-list>
             </v-sheet>
